re order file per logic for avoid using only one long file
